Changes:
Change shared_sanity_checks to _sanity_checks and move to end of file.
Change condition of sanity checks on parameters.  Previous setup to trigger tests on parameter "a" was of the form"if params.get("a", None) is not None", which means tests would not be triggered if params["a"] exists but is None.  New setup is "if 'a' in params: a = params.get("a", None)"

Change migration model tests to check all values rather than only a random sampling of values

Add a test case for radiation and Stouffer models to ensure correct functioning when there are non-unique, non-zero elements in a row of the distance matrix.

Fixed a comment for clarity.