Update requirements.txt
removed laser-core and the link --index-url https://packages.idmod.org/artifactory/api/pypi/pypi-production/simple  from the requirements.txt file. Wanted the notebooks to build in the GH pages docs, but there's an issue with the deploy step in "preparing metadata (pyproject.toml)" step that causes it to fail. So removing these and just having the notebooks not build until I can troubleshoot the issue.